feat(auth): base login fields on GoTrue providers

# Conflicts:
#	app/(auth)/login/login-client.tsx
#	app/(auth)/register/page.tsx
