+++ app/(auth)/login/page.tsx
@@ -2,6 +2,7 @@ import { Suspense } from 'react'
 import { cookies } from 'next/headers'
 import { AuthPageSkeleton } from '@/components/auth/AuthPageSkeleton'
 import { LOGIN_METHOD_COOKIE, isLoginMethod } from '@/lib/auth/login-method'
+import { fetchAuthSettings, type GoTrueAuthSettings } from '@/lib/auth/gotrue-providers'
 import { LoginClient } from './login-client'
 
 // Server component: reads the method-hint cookie so the panel opens in the
@@ -12,9 +13,16 @@ export default async function LoginPage() {
   const cookieStore = await cookies()
   const stored = cookieStore.get(LOGIN_METHOD_COOKIE)?.value
 
+  // Fetch auth settings from GoTrue: enabled providers, password login,
+  // and registration status. Lightweight GET, cached for 60 seconds.
+  const authSettings: GoTrueAuthSettings = await fetchAuthSettings()
+
   return (
     <Suspense fallback={<AuthPageSkeleton />}>
-      <LoginClient initialMethod={isLoginMethod(stored) ? stored : null} />
+      <LoginClient
+        initialMethod={isLoginMethod(stored) ? stored : null}
+        authSettings={authSettings}
+      />
     </Suspense>
   )
 }
