+++ components/auth/OAuthButton.tsx
@@ -4,31 +4,46 @@ import { useState } from 'react'
 import { useLocale, useTranslations } from 'next-intl'
 import { createClient } from '@/lib/supabase/client'
 import { Button } from '@/components/ui/button'
-import { Loader2 } from 'lucide-react'
+import { Loader2, KeyRound } from 'lucide-react'
 import { getErrorMessage, type ErrorLocale } from '@/lib/errors/get-error-message'
 import { GoogleMark } from '@/components/ui/provider-marks'
-
+import type { ResolvedProvider } from '@/lib/auth/gotrue-providers'
 
 /**
- * "Continue with Google" for the login and register pages.
- *
- * Kicks off the Supabase OAuth redirect; the round-trip lands in
- * /auth/callback (PKCE code exchange), which owns MFA routing, invite
- * acceptance and silent-team creation for OAuth sign-ins and sign-ups alike.
- * The flow=oauth marker lets the callback tag failures so the login page
- * shows Google-specific copy instead of the email-confirmation framing.
+ * Render the brand mark for a known provider, or a generic key icon for
+ * custom OIDC providers.
  */
-export function GoogleAuthButton({
+function ProviderMark({ provider }: { provider: ResolvedProvider }) {
+  if (provider.id === 'google') return <GoogleMark />
+  // Known providers without a dedicated mark show the first letter
+  if (!provider.isCustom) {
+    return (
+      <span className="flex h-4 w-4 items-center justify-center text-[11px] font-bold text-muted-foreground">
+        {provider.label.charAt(0)}
+      </span>
+    )
+  }
+  // Custom OIDC providers get a generic key icon
+  return <KeyRound className="h-4 w-4 text-muted-foreground" />
+}
+
+/**
+ * Generic OAuth login button. Works with any Supabase GoTrue provider.
+ *
+ * For known providers (Google, GitHub, etc.) the button shows the brand
+ * name; for custom OIDC providers it shows "Sign in with SSO" style text.
+ *
+ * Kicks off the Supabase OAuth redirect, with flow=oauth so
+ * /auth/callback can tag failures.
+ */
+export function OAuthButton({
+  provider,
   onError,
   compact = false,
   next,
 }: {
+  provider: ResolvedProvider
   onError: (message: string) => void
-  /**
-   * Half-width alternative-method chip on the login panel: shows just the
-   * mark and "Google" (a brand name, never translated), with the full label
-   * kept as the accessible name.
-   */
   compact?: boolean
   /**
    * Post-auth destination, already passed through safeReturnTo by the caller.
@@ -51,7 +66,7 @@ export function GoogleAuthButton({
       callback.searchParams.set('flow', 'oauth')
       if (next && next !== '/') callback.searchParams.set('next', next)
       const { error } = await supabase.auth.signInWithOAuth({
-        provider: 'google',
+        provider: provider.id as Parameters<typeof supabase.auth.signInWithOAuth>[0]['provider'],
         options: {
           redirectTo: callback.toString(),
         },
@@ -60,12 +75,15 @@ export function GoogleAuthButton({
         onError(getErrorMessage(error, { context: 'auth', locale: errorLocale }))
         setIsRedirecting(false)
       }
-      // On success the browser navigates away; keep the spinner until then.
     } catch (error) {
       onError(getErrorMessage(error, { context: 'auth', locale: errorLocale }))
       setIsRedirecting(false)
     }
   }
+
+  const label = provider.isCustom
+    ? tAuth('continue_with_sso')
+    : tAuth('continue_with_provider', { provider: provider.label })
 
   return (
     <Button
@@ -74,16 +92,16 @@ export function GoogleAuthButton({
       className={compact ? 'h-10 w-full gap-2' : 'w-full h-11'}
       onClick={handleClick}
       disabled={isRedirecting}
-      aria-label={tAuth('continue_with_google')}
+      aria-label={label}
     >
       {isRedirecting ? (
         <Loader2 className={compact ? 'h-4 w-4 animate-spin' : 'mr-2 h-4 w-4 animate-spin'} />
       ) : (
         <span className={compact ? 'flex items-center' : 'mr-2 flex items-center'}>
-          <GoogleMark />
+          <ProviderMark provider={provider} />
         </span>
       )}
-      {compact ? 'Google' : tAuth('continue_with_google')}
+      {compact ? provider.label : label}
     </Button>
   )
 }
