+++ messages/en.json
@@ -302,7 +302,8 @@
     "login_error_rate_limited": "Too many sign-in attempts. Wait a moment and try again.",
     "login_error_user_banned": "This account has been suspended. Contact support if you think this is a mistake.",
     "login_error_reset_link": "Reset your password",
-    "continue_with_google": "Continue with Google",
+    "continue_with_provider": "Continue with {{provider}}",
+    "continue_with_sso": "Continue with SSO",
     "callback_error_title_oauth": "Google sign-in didn't work",
     "callback_error_body_oauth": "The Google sign-in could not be completed. Try again, or sign in with email and password.",
     "session_idle": "You were inactive. Sign in again.",
