+++ messages/sv.json
@@ -302,7 +302,8 @@
     "login_error_rate_limited": "För många inloggningsförsök. Vänta en stund och försök igen.",
     "login_error_user_banned": "Kontot är avstängt. Kontakta supporten om du tror att det är fel.",
     "login_error_reset_link": "Återställ lösenordet",
-    "continue_with_google": "Fortsätt med Google",
+    "continue_with_provider": "Fortsätt med {{provider}}",
+    "continue_with_sso": "Fortsätt med SSO",
     "callback_error_title_oauth": "Google-inloggningen fungerade inte",
     "callback_error_body_oauth": "Inloggningen med Google kunde inte slutföras. Försök igen, eller logga in med e-post och lösenord.",
     "session_idle": "Du har varit inaktiv. Logga in igen.",
